Refactors error handling for OpenAI endpoints
Consolidates error response creation logic into the base class.

This change refactors the error handling mechanism for OpenAI-compatible endpoints. It removes duplicated error response creation functions from `utils.py` and the handlers, and instead, implements centralized `create_error_response` and `create_streaming_error_response` methods in the `OpenAIServingBase` class. The handlers are updated to use these base class methods. This promotes code reuse and ensures consistency in error response formatting across all OpenAI endpoints. Also updates the streaming response to properly format the data.
